[nrf toup] platform: nordic_nrf: Add option to initialize hardware to reset on boot
Add an option to the nordic platforms to initialize the hardware state
to the its reset values when the TF-M firmware is booting.
The option is useful when the TF-M firmware image is chain-loaded, for
example, a bootloader, and we need to guarantee that the internal
states of the architecture core blocks are restored to the reset values.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>