platform: common: Update common linker script to support Nordic platform
Update the common linker script to support Nordic platform specific
requirements.
Since the Nordic platform has the Nordic specific peripheral SPU instead
of the ARM provided peripheral SAU the veneers location needs to be
different.
The Nordic SPU supports putting the Veneers at the end of an SPU Flash
page, which is 16 KB on nrf5340 and 32 KB on nrf9160.
The Nordic platform has support for a NULL pointer detection, which
requires that executable code be put outside of the first MPU region.

Change-Id: Icc044f8c634fed2fe0013a4176c226c3101d2d27
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>