platform: lairdconnectivity: Use the common linker files
Use the common linker scripts for the lairdconnectivity platform.
The nordic linker file for secure has become out of sync with the
changes made to the linking and enabling of HEAP, resulting in
compilation errors when using isolation level 2.
This fixes this issue by using the common linker file instead.

Change-Id: Id3e14d8b8b79e3e75a5806750e7a000803a8ea88
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>