platform: nordic_nrf: Header file for boundary handle macros
Add header file for boundary handle macros.
This avoids having to redefine these defines in a source file that
needs these defines for memory access checks.
The mmio_defs.h header cannot be included because it defines a
global struct and can only be included in one source file in the build.

Change-Id: Iad257825c5a5b2f2c08897b66961e1baad891747
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>