CC3XX: chachapoly_alt: Improve check for NULL pointer
Add missing check for ctx object being NULL.
Remove the check for key == NULL which will then always false.
Provide the context object to the helper function
chachapoly_crypt_and_tag which makes it obvious why the key object
pointer can never be NULL as this is an array inside the context object.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>