CC3XX: entropy: Support smaller random bytes requests
Add support in mbedtls_hardware_poll for requesting random number
output to buffer less than the buffer size returned by
LLF_RND_GetTrngSource.
Since mbedtls always requests a fixed buffer size this is not currently
an issue, but this is a strange limitation of the function.
Improve readability by not giving the ignored len field to
LLF_RND_GetTrngSource.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>