crypto: Improve PSA interface config file include
Remove the include of the config file in the middle of the crypto_sizes
header. This was additionaly guarded by a config this expected to be
set from a configuration file. So this include was only included if the
config was included, making it redundant and useless.

Move the only include from crypto_types header.
It is only needed in crypto_sizes at the moment, and finding it in a
different header other than where it is needed is very unexpected and
hard to find out when debugging. Also this relies on the order of
header includes which seems like an unneccesey dependency to have.

Move it to the main PSA header instead.
In mbedtls it is included in every header except for the for the main
PSA header. But instead of doing the same having it in the main
header should be enough.

Change-Id: I953fe44a68788dc926a4f726a0408767886c5146
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>