crypto: Fix PSA max hash size without configuration header included
The PSA header only includes the configuration header when
MBEDTLS_PSA_CRYPTO_CONFIG_FILE is given on the command line.
If the configuration file has not been included assume that the
maximum hash size is the maximum possible instead of the minimal
possible.
This is consistent with how PSA_VENDOR_ECC_MAX_CURVE_BITS is defined.

Change-Id: I4c137a947ca0dbdbddf7eaac74e6ce59ae988224
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>